openimage full screen ,add qtwidget

+++ src/CGRenderView/CGCamera.h
@@ -40,7 +40,7 @@ namespace CGRender
 		// camera options
 		float MovementSpeed;
 		float MouseSensitivity;
-		float Zoom;
+		double Zoom;
 
 		CGCamera(glm::vec3 position = glm::vec3(0.0f, 0.0f, 1000.0f), glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f), float yaw = YAW, float pitch = PITCH);
 		//CGCamera(float posX, float posY, float posZ, float upX, float upY, float upZ, float yaw, float pitch);
